+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 0.11.0 
+
+* ...
+
 # Version 0.10.0
 
 ## New features